# docs/api/UNIFIED_API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: AIVillage Unified API Gateway
  description: |
    Production-ready unified API for AIVillage services including:
    - Agent Forge 7-phase training pipeline
    - P2P/Fog computing integration
    - JWT authentication with MFA support
    - Real-time WebSocket updates
    - Comprehensive health monitoring
  version: 1.0.0
  contact:
    name: AIVillage API Support
    url: https://github.com/DNYoussef/AIVillage
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.aivillage.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /v1/models:
    get:
      summary: List Models
      description: List all trained models for authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                success: true
                data:
                  models: []
                  total_count: 0
                message: "Models retrieved successfully"
  /v1/chat:
    post:
      summary: Chat with Model
      description: Send message to trained model
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            example:
              model_id: "cognate-model-123"
              message: "Hello, how can you help me?"
              conversation_id: "conv-456"
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              example:
                success: true
                data:
                  model_id: "cognate-model-123"
                  response: "Hello! I'm a production-trained AI model..."
                  conversation_id: "conv-456"
                message: "Chat completed successfully"

  # P2P/Fog Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          description: Response data
        message:
          type: string
          description: Human-readable message
        error_code:
          type: string
          nullable: true
          description: Error code if applicable
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        request_id:
          type: string
          nullable: true
          description: Unique request identifier
...
    ChatRequest:
      type: object
      required:
        - model_id
        - message
      properties:
        model_id:
          type: string
          description: Model ID to chat with
        message:
          type: string
          minLength: 1
          maxLength: 5000
          description: Message to send to model
        conversation_id:
          type: string
          nullable: true
          description: Optional conversation ID
        stream:
          type: boolean
          default: false
          description: Whether to stream response
...